+++ openrem/remapp/views_admin.py
@@ -1324,6 +1324,9 @@ def chart_options_view(request):
             user_profile.plotRemoveCategoryTrailingWhitespace = general_form.cleaned_data[
                 "plotRemoveCategoryTrailingWhitespace"
             ]
+            user_profile.plotRemoveCategoryMultipleWhitespace = general_form.cleaned_data[
+                "plotRemoveCategoryMultipleWhitespace"
+            ]
 
             if "mean" in general_form.cleaned_data["plotAverageChoice"]:
                 user_profile.plotMean = True
@@ -1525,6 +1528,7 @@ def chart_options_view(request):
         "plotHistogramGlobalBins": user_profile.plotHistogramGlobalBins,
         "plotCaseInsensitiveCategories": user_profile.plotCaseInsensitiveCategories,
         "plotRemoveCategoryTrailingWhitespace": user_profile.plotRemoveCategoryTrailingWhitespace,
+        "plotRemoveCategoryMultipleWhitespace": user_profile.plotRemoveCategoryMultipleWhitespace,
         "plotThemeChoice": user_profile.plotThemeChoice,
         "plotColourMapChoice": user_profile.plotColourMapChoice,
         "plotFacetColWrapVal": user_profile.plotFacetColWrapVal,