+++ openrem/remapp/views_charts_rf.py
@@ -282,6 +282,7 @@ def rf_plot_calculations(f, user_profile, return_as_dict=False):
         fields,
         data_point_name_lowercase=user_profile.plotCaseInsensitiveCategories,
         data_point_name_remove_trailing_whitespace=user_profile.plotRemoveCategoryTrailingWhitespace,
+        data_point_name_remove_multiple_whitespace=user_profile.plotRemoveCategoryMultipleWhitespace,
         data_point_value_multipliers=value_multipliers,
         uid="pk",
     )